LLM_evaluations/requirements-apis.txt
anthropic==0.40.0
openai==1.76.1
google-generativeai==0.8.5
tenacity==9.1.2
tqdm==4.66.4